+++ libs/langgraph-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @langchain/langgraph-cli
 
+## 1.4.6-rc.0
+
+### Patch Changes
+
+- Updated dependencies [[`0a9611a`](https://github.com/langchain-ai/langgraphjs/commit/0a9611af927bd7933e6eb96161acd86584f0d4b0)]:
+  - @langchain/langgraph-api@1.4.6-rc.0
+
+## 1.4.5
+
+### Patch Changes
+
+- Updated dependencies [[`a2a59ec`](https://github.com/langchain-ai/langgraphjs/commit/a2a59ec6f8fdd93d4520d86fceab8a234dacf978)]:
+  - @langchain/langgraph-api@1.4.5
+
 ## 1.4.4
 
 ### Patch Changes